+++ client/webpack/src/router/matricula.js
@@ -1,0 +1,14 @@
+// Paginas
+import matriculados from '@/pages/matricula/matriculados'
+import como_matricularse from '@/pages/matricula/como_matricularse'
+
+export default [
+  {
+    path: '/lista-de-matriculados',
+    component: matriculados
+  },
+  {
+    path: '/como-matricularse',
+    component: como_matricularse
+  },
+]
